pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cmt"
version = "0.1.0"
description = "Content Moderation Thesis - Evaluation and Training Pipeline"
authors = [
  { name = "Thesis Student", email = "student@hs-flensburg.de" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "pandas",
    "Pillow",
    "open_clip_torch",
    "timm",
    "scikit-learn",
    "pyyaml",
    "tqdm",
    "matplotlib",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["cmt*"]
